refactor: restructure COP-1 into a proper module
Pull as much as possible out of __init__.py and into submodules
